requirements.txt
uvicorn==0.23.2
fastapi==0.101.1
PyJWT==2.8.0
SQLAlchemy==2.0.20